move sensor_itf to driver_itf
