+++ hw/driver/src/driver.c
@@ -1,0 +1,72 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+#include <string.h>
+#include <errno.h>
+#include <assert.h>
+#include "os/mynewt.h"
+#include "driver/driver.h"
+
+/**
+ * Lock access to the driver_itf specified by si. Blocks until lock acquired.
+ *
+ * @param The driver_itf to lock
+ * @param The timeout
+ *
+ * @return 0 on success, non-zero on failure.
+ */
+int
+driver_itf_lock(struct driver_itf *si, uint32_t timeout)
+{
+    int rc;
+    os_time_t ticks;
+
+    if (!si->si_lock) {
+        return 0;
+    }
+
+    rc = os_time_ms_to_ticks(timeout, &ticks);
+    if (rc) {
+        return rc;
+    }
+
+    rc = os_mutex_pend(si->si_lock, ticks);
+    if (rc == 0 || rc == OS_NOT_STARTED) {
+        return (0);
+    }
+
+    return (rc);
+}
+
+/**
+ * Unlock access to the driver_itf specified by si.
+ *
+ * @param The driver_itf to unlock access to
+ *
+ * @return 0 on success, non-zero on failure.
+ */
+void
+driver_itf_unlock(struct driver_itf *si)
+{
+    if (!si->si_lock) {
+        return;
+    }
+
+    os_mutex_release(si->si_lock);
+}
