+++ hw/sensor/src/sensor.c
@@ -852,54 +852,6 @@ sensor_pkg_init(void)
 #endif
 }
 
-/**
- * Lock access to the sensor_itf specified by si. Blocks until lock acquired.
- *
- * @param The sensor_itf to lock
- * @param The timeout
- *
- * @return 0 on success, non-zero on failure.
- */
-int
-sensor_itf_lock(struct sensor_itf *si, uint32_t timeout)
-{
-    int rc;
-    os_time_t ticks;
-
-    if (!si->si_lock) {
-        return 0;
-    }
-
-    rc = os_time_ms_to_ticks(timeout, &ticks);
-    if (rc) {
-        return rc;
-    }
-
-    rc = os_mutex_pend(si->si_lock, ticks);
-    if (rc == 0 || rc == OS_NOT_STARTED) {
-        return (0);
-    }
-
-    return (rc);
-}
-
-/**
- * Unlock access to the sensor_itf specified by si.
- *
- * @param The sensor_itf to unlock access to
- *
- * @return 0 on success, non-zero on failure.
- */
-void
-sensor_itf_unlock(struct sensor_itf *si)
-{
-    if (!si->si_lock) {
-        return;
-    }
-
-    os_mutex_release(si->si_lock);
-}
-
 
 /**
  * Lock access to the sensor specified by sensor.  Blocks until lock acquired.
